backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
black==25.9.0
celery==5.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.11.0
distro==1.9.0
Django==5.2.7
django-simple-history==3.10.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
factory_boy==3.3.3
Faker==37.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mypy_extensions==1.1.0
openai==2.6.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytokens==0.2.0
PyYAML==6.0.3
qrcode==8.2
redis==5.0.7
referencing==0.37.0
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
ruff==0.14.1
sentry-sdk==2.42.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==13.0.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
